fix(rewrite): rewrite question mark
This is to make resources like `font-awesome`, to work when there is
query param question mark symbol in the given URL. When one uses the
browser to download the file, this issue does not occur since the file
browser will automatically convert question mark into its escaped form.

Note that static files with hex / anchor link would not work, since hex
is a purely client thing and never made it to NGINX. This is
unresolvable but should be okay since these resources are usually
backups of other font URLs that are non-anchor based.

The set-up now switches to NGINX since it can perform rewrite more
easily.